+++ src/config/configTransaction.js
@@ -3,9 +3,36 @@
 /////////////////////////////////////////////////////
 
 // A presets of supported transaction configurations
+//
+// Note: With first iteration of hosted configs, we are unlikely to support
+//       multiple transaction types, even though this template has some
+//       rudimentary support for it.
+
+/**
+ * Configuration options for transaction experience:
+ * - type:            Unique string. This will be saved to listing's public data on
+ *                    EditListingWizard.
+ * - label            Label for the transaction type. Used as microcopy for options to select
+ *                    transaction type in EditListingWizard.
+ * - process          Transaction process. This will be saved to listing's public data
+ *                    (together with alias) as transctionProcessAlias.
+ *                    The process must match one of the processes that this client app can handle
+ *                    (check src/util/transaction.js) and the process must also exists in correct
+ *                    marketplace environment.
+ * - alias            Valid alias for the aforementioned process.
+ * - unitType         Unit type is mainly used as pricing unit. This will be saved to
+ *                    listing's public data.
+ *                    Recommendation: don't use same unit types in completely different processes
+ *                    ('item' sold should not be priced the same as 'item' booked).
+ * - showStock        This is relevant only to listings with product transactions.
+ *                    If set to false, stock management is not showed and the listing is
+ *                    considered unique (stock = 1).
+ *                    Default: true.
+ */
+
 export const transactionTypes = [
   {
-    type: 'sell-bicycles',
+    type: 'product-selling',
     label: 'Sell bicycles',
     process: 'flex-product-default-process',
     alias: 'release-1',
@@ -13,31 +40,33 @@ export const transactionTypes = [
     showStock: false,
   },
   {
-    type: 'rent-bicycles-daily',
-    label: 'Rent bicycles: daily',
+    type: 'daily-booking',
+    label: 'Daily booking',
     process: 'flex-booking-default-process',
     alias: 'release-1',
     unitType: 'day',
   },
   {
-    type: 'rent-bicycles-nightly',
-    label: 'Rent bicycles: nightly',
+    type: 'nightly-booking',
+    label: 'Nightly booking',
     process: 'flex-booking-default-process',
     alias: 'release-1',
     unitType: 'night',
   },
   {
-    type: 'rent-bicycles-hourly',
-    label: 'Rent bicycles: hourly',
+    type: 'hourly-booking',
+    label: 'Hourly booking',
     process: 'flex-booking-default-process',
     alias: 'release-1',
     unitType: 'hour',
   },
 ];
 
-// SearchPage can enforce searching of listings with valid transactionType, process and unitType
+// SearchPage can enforce listing query to only those listings with valid transactionType
 // However, it only works if you have set 'enum' type search schema for the public data fields
 //   - transactionType
+//
+//  Similar setup could be expanded to 2 other extended data fields:
 //   - transactionProcessAlias
 //   - unitType
 //